Ignore invalid xml chars... from https://github.com/RavnSystems/ews-java-api
See this commit:
RavnSystems@845f20f

and this issue:
OfficeDev#353

Looks like exchange does not properly excape some XML chars
when present in email. There seems to be a way to allow unicode
by forcing "Xml 1.1" instead of "Xml 1.0" but this is reported
to not work in some versions of exchange.